pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "http-request-randomizer>=1.3.2",
    "lxml>=5.4.0",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "plyer>=2.1.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-socketio[async-client,asyncio-client,client]>=5.13.0",
    "requestium>=0.4.0",
    "requests>=2.32.3",
    "retry>=0.9.2",
    "selenium>=4.33.0",
    "sox>=1.5.0",
    "stem>=1.8.2",
    "tbselenium>=0.9.0",
    "tqdm>=4.67.1",
    "ytmusicapi>=1.10.3",
]
